Skip SCHEDULED-timestamps expressed as diary-sexps for now
NOTE: it is obviously very bad to skip any SCHEDULED/DEADLINE
timestamps.  Could lead to surprises!  So we need to implement correct
parsing soon.

Skipping for now so that so that a file can be parsed in full and
org-mem is more usable.  Want to implement a warning like "diary-sexps
are not supported and org-mem will not list such entries as having any
SCHEDULED/DEADLINE".  But it's not obvious how to implement that
warning either, so aiming to save effort and just work on the parser.

https://github.com/meedstrom/org-mem/issues/21