Refactor file name discovery (see commit message)
Removed hack: The fact `org-mem-initial-scan-hook' could run twice (this
was documented).  Now it runs once, so you don't need to think about that.

Removed hack: The function `org-mem--tramp-file-p'.  It could theoretically
cause Tramp files to be included in the files to scan, and mandate that
something load Tramp and then reset org-mem in order to sanitize.  (No idea if
that was ever likely.)

Removed hack: The boolean `org-mem--first-run'.

Org-mem is better at finding true names ahead-of-time, especially at init
time, instead of waiting for parser to discover a symlink and trigger a
re-scan.  Now it probably only happens later on, as files are renamed during
normal usage.

Org-mem will now find real true names of files mentioned in org-id-locations
where some parent directory is actually a symlink.

Changed org-mem--dir-files-recursive, now accepts a list of suffixes.