Treat files with blank #+TITLE as if they have no such line
This permits `org-node-rename-file-by-title' to fall back
on the first heading.

And some people may routinely create files with a template that adds
blank keyword lines like #+title and #+date even if they do not always
fill them, in which case there can be a lot of files with the title "".